z3-sys: Use generated enum value files via bindgen.
This lets us keep our nice documentation, but also make it easier
to be sure that we're in sync with the C headers with respect to
the various values, what's been removed and what's been added.

If a value is removed from the C side, then it will no longer be
in the generated files and we'll get a compile failure.

If a value is added on the C side, then we'll get a compile warning
about the unused variant of the enum and we can then update our side.

The `scripts/generate_enums.sh` script will update all of the
generated enum files. It needs to be provided with the path to the
`z3.h` header.