z3-sys/build: fix link order for static lib
This tweaks link order when building a static libz3, so that
dangling C++ stdlib symbols can be resolved by the library appearing
aftwerward in linking order.
It fixes a linking failure when trying to use latest Z3 from git,
as it would otherwise fail to resolve C++ stdlib symbols used
by the library.