ci: Use slightly older emscripten for now
An update to emscripten brought a newer compiler and that found
a longstanding bug in Z3 that broke compilation.

For now, use a slightly older version until a new version of
Z3 (after 4.13.0) is released with the fix.

See https://github.com/Z3Prover/z3/pull/7235 for the Z3 fix.