feat!: trait-based conversions and operations (#410)
This PR introduces the traits IntoAst and IntoAstFromCtx, which capture the ability of a type to be expressed in a Z3 ast, and changes the targets of most operations over all Ast types to take IntoAst<Self>. It additionally provides blanket implementations for existing Ast types and implementations on the primitive number and string types for BV, Int, Float, Real, and String.

The consequence of this is that you can still construct expressions like before, but now most expressions can be simplified to avoid the boilerplate of manually constructing trivial constants. Instead, you can use primitive rust types in most places you feel like you should be able to when building expressions: e.g. "bv + 4"